Some UI Logic has been changed.

+++ OrderProcessingApp.Models/Order.cs
@@ -4,7 +4,7 @@
     {
         public string orderId { get; set; }
 
-        public int OrderTotal { get; set; }
+        public decimal OrderTotal { get; set; }
 
         public string CustomerType { get; set; }
     }
